Initialize Website Rebuild
First commit of rebuilding project for a Puppy Training website.

This repo includes the original website "Puppy Ph.D. Classes.html" and
all its dependancies in "Puppy Ph.D. Classes_files". My work is
contained in "index.html", "styles.css", and "images" which is a direct
copy of the dependancies folder from the source. The goal of this
project is to recreate the website entirely with the same basic style
while improving the functionality and spacing. Additionally, I'd like
to improve the style to be more legible and visually present.

First edits include adding the boilerplate, headers, a background image
and the images for the puppy wearing the grad cap.

TO-DO:
    - Organize the whole website into folders with appropriate resources
    and webpages.
    - Recreate website to be exact same layout as before
    - Improve style...
    - I'm sure there will be more as I go along.