Big changes: images, links, styling
This commit is the combo of two major working sessions. I had to step
away for a while and forgot that I was having issues pushing to github.

The first commit included adding images, changing the buttons to links,
and creating the framework for the other pages. making the buttons links
looks a little less clean in my opinion, however it's more consistant
with the original site's style.

The second commit involves beginning styling with css and flex for the
header. I've made significant progress in organizing he header, however
there's more to come!