Structure page and style
This commit includes many edits to the div trees that organize the
content for easy flex positioning. During this process, I'm finding that
a simpler page creates a more user-friendly interface already. Many
classes and ids were added add the styles.css was organized better to
follow the html content.